+++ examples/websocket_api/account/OrderRateLimits/OrderRateLimits.go
@@ -1,0 +1,33 @@
+package main
+
+import (
+	"context"
+	"fmt"
+	"log"
+
+	binance_connector "github.com/binance/binance-connector-go"
+)
+
+func main() {
+	OrderRateLimits()
+}
+
+func OrderRateLimits() {
+	client := binance_connector.NewWebsocketAPIClient("api_key", "secret_key")
+	err := client.Connect()
+	if err != nil {
+		log.Printf("Error: %v", err)
+		return
+	}
+	defer client.Close()
+
+	response, err := client.NewAccountOrderRateLimitsService().Do(context.Background())
+	if err != nil {
+		log.Printf("Error: %v", err)
+		return
+	}
+
+	fmt.Println(binance_connector.PrettyPrint(response))
+
+	client.WaitForCloseSignal()
+}